quick

+++ 4_fetchcontent/build.sh
@@ -2,7 +2,8 @@
 set -euo pipefail
 
 BUILD_DIR=${BUILD_DIR:-build}    # if BUILD_DIR is not set as environment variable, default build directory is "build"
-rm -rf $BUILD_DIR
+[ -d $BUILD_DIR ] && rm -rf $BUILD_DIR
+
 
 # -S .                : sets source directory to the current directory
 # -B build            : sets build directory to the build folder (creates if it does not exist)
@@ -12,4 +13,3 @@ rm -rf $BUILD_DIR
 cmake -S . -B "${BUILD_DIR}" -G "Unix Makefiles"
 cmake --build "${BUILD_DIR}" --parallel
 
-# ctest --rerun-failed --output-on-failure
